style: move class, interface - validationRule and implmenet classes
