--- src/main/java/co/fineants/api/domain/member/domain/entity/ValidationRule.java
@@ -1,7 +1,0 @@
-package co.fineants.api.domain.member.domain.entity;
-
-public interface ValidationRule {
-	void validate(String text);
-
-	void validate(Member member);
-}
