+++ src/main/java/co/fineants/api/domain/member/domain/rule/EmailFormatRule.java
@@ -1,9 +1,10 @@
-package co.fineants.api.domain.member.domain.entity;
+package co.fineants.api.domain.member.domain.rule;
 
 import java.util.regex.Pattern;
 
 import org.apache.logging.log4j.util.Strings;
 
+import co.fineants.api.domain.member.domain.entity.Member;
 import co.fineants.api.global.errors.exception.business.EmailInvalidInputException;
 
 public class EmailFormatRule implements ValidationRule {
