+++ src/main/java/co/fineants/api/domain/member/domain/rule/ValidationRule.java
@@ -1,0 +1,9 @@
+package co.fineants.api.domain.member.domain.rule;
+
+import co.fineants.api.domain.member.domain.entity.Member;
+
+public interface ValidationRule {
+	void validate(String text);
+
+	void validate(Member member);
+}
