+++ src/main/java/co/fineants/api/domain/member/service/MemberService.java
@@ -28,14 +28,14 @@ import co.fineants.api.domain.member.domain.dto.request.VerifyEmailRequest;
 import co.fineants.api.domain.member.domain.dto.response.ProfileChangeResponse;
 import co.fineants.api.domain.member.domain.dto.response.ProfileResponse;
 import co.fineants.api.domain.member.domain.dto.response.SignUpServiceResponse;
-import co.fineants.api.domain.member.domain.entity.EmailDuplicationRule;
-import co.fineants.api.domain.member.domain.entity.EmailFormatRule;
 import co.fineants.api.domain.member.domain.entity.Member;
 import co.fineants.api.domain.member.domain.entity.MemberRole;
-import co.fineants.api.domain.member.domain.entity.NicknameDuplicationRule;
-import co.fineants.api.domain.member.domain.entity.NicknameFormatRule;
 import co.fineants.api.domain.member.domain.entity.Role;
-import co.fineants.api.domain.member.domain.entity.ValidationRule;
+import co.fineants.api.domain.member.domain.rule.EmailDuplicationRule;
+import co.fineants.api.domain.member.domain.rule.EmailFormatRule;
+import co.fineants.api.domain.member.domain.rule.NicknameDuplicationRule;
+import co.fineants.api.domain.member.domain.rule.NicknameFormatRule;
+import co.fineants.api.domain.member.domain.rule.ValidationRule;
 import co.fineants.api.domain.member.repository.MemberRepository;
 import co.fineants.api.domain.member.repository.MemberRoleRepository;
 import co.fineants.api.domain.member.repository.RoleRepository;
